Change scc_default_options to function

+++ tests/stress_nng_clustering.c
@@ -64,7 +64,7 @@ static void iscc_make_batch_options(scc_ClusterOptions* out_options,
 }
 
 static scc_ClusterOptions iscc_translate_options(const uint32_t size_constraint,
-                                                 const uintmax_t num_types,
+                                                 const uint32_t num_types,
                                                  const uint32_t* const type_constraints,
                                                  const size_t len_type_labels,
                                                  const scc_TypeLabel* const type_labels,
