+++ tests/test_nng_clustering.c
@@ -49,7 +49,7 @@ void iscc_run_type_nonval_tests(scc_SeedMethod seed_method,
                                 double secondary_radius);
 
 static scc_ClusterOptions iscc_translate_options(const uint32_t size_constraint,
-                                                 const uintmax_t num_types,
+                                                 const uint32_t num_types,
                                                  const uint32_t* const type_constraints,
                                                  const size_t len_type_labels,
                                                  const scc_TypeLabel* const type_labels,
